os-practice: use predefined memory functions and define build target
Makes sure we don't have to reimplement standard memory functions
(e.g. "memset") using the "#[no_mangle]" attribute. Saves the headache
of having to deal with the possibility of an incorrect implementation
that would lead to undefined behavior. Also sets a default build target
so we no longer have to point to the json every time

Signed-off-by: Ryan Sullivan <rysulliv@redhat.com>