os-practice: implement breakpoint exception handler and restore scratch registers
Now the wrapper functions the handler! macros create will store then
restore all of the scratch registers before returning to the caller.
This allows us to return to where the exception occurred, in later
handler functions we can implement fixes corresponding to the interrupt
that need to occur before returning. Currently only the breakpoint
handler returns as it doesn't require any fixing, plus, on x86 hardware
the return instruction pointer points to the instruction following the
int3 anyway as well so no need to check that either

Signed-off-by: Ryan Sullivan <ryanbakersullivan@gmail.com>