os-practice: added in more serial logging to try and determine error
Currently we are loading the TSS with the GDT and pointing to a stack
created in the TSS's IST with our double_fault_handler()'s stack_idx,
but when a kernel stack overflow occurs, the system does not seem to
switch to the created stack. I added in more logging of kernel structs
but still cannot seem to find where the disconnect is happening?

TODO: get GDB working with the kernel and track down where the triple
fault occurs

Signed-off-by: Ryan Sullivan <rysulliv@redhat.com>