os-practice: fix IST stack switching troubles
Thanks to u/mpetch who pointed out the error, when setting the IST index
in the IDT entry options you need to add 1 as a value of 0 in the IST
index bits is seen as DISABLED and for any entry !=0 then the IST idx is
just the value of those bits minus one. So when calling
set_stack_idx(DOUBLE_FAULT_IST_IDX) the EntryOptions for the
double_fault_handler() either never enabled it for a value of 0 or
pointed to DOUBLE_FAULT_IST_IDX-1 for any other value, leading to a page
fault when stack switching, leading to a triple fault and a reboot.

Signed-off-by: Ryan B. Sullivan <ryanbakersullivan@gmail.com>